distinguish between method call and method response
We can not only use the fact that we have a method call inflight with
the observed id. Because the language server may has called a method
on use with the same id.

Therefore we have to also check if the incoming message contains
a method field.

method + id -> method call
id -> method response
method -> notification