fix diagnostic highlighting and message
The way it is currently set up, we have a defined range that is used to
apply a diagnostic colour over in cases where there are errors. That
range is one character longer than it needs to be (i.e resulting in
highlighting whitespace or what ever it could be). Some LSPs reports
diagnostics where it starts and ends at the same line and character
resulting in no highlighting at all.

`textDocument/publishDiagnostics` doesn't always have a `code` field,
resulting in `diagnostic.code` being nil and so it is replaced with
`error` in those cases.