Change vis_sel_to_pos to count lines itself
This avoids the screen scrolling when receiving diagnostics from the
server, which currently means that saving the file randomly centers the
cursor.

This is a first version, I think that an optimisation is possible where
it doesn't loop through the file for every selection, but loops once,
converting every selection passed.